+++ apps/client/src/views/posts/ui/index.tsx
